feat: Cleared the merge conflicts
