+++ app/src/main/java/com/sq/thed_ck_licker/player/Player.kt
@@ -12,9 +12,9 @@ import androidx.compose.ui.unit.dp
 
 @Composable
 fun HealthBar(health: Float, modifier: Modifier = Modifier) {
-    println(health)
-    println (health % 100)
-    println ((health % 100) / 100f)
+    println("health: $health")
+    println("health % 100: ${health % 100}")
+    println("health / 100: ${(health % 100) / 100f}")
     Column(modifier = modifier.fillMaxWidth()) {
         Text(
             //TODO This kind a should be generic some how
@@ -24,22 +24,21 @@ fun HealthBar(health: Float, modifier: Modifier = Modifier) {
         //we might want to look into some other way of doing the hp bar in the future :D
         //this way the bar starts off red and goes from left to right
         LinearProgressIndicator(
-            progress = { if(health == 100f) 0f else (1 -(health % 100) / 100f)},
+            progress = { if (health == 100f) 0f else (1 - (health % 100) / 100f) },
             modifier = modifier
                 .fillMaxWidth()
                 .height(15.dp),
-            color =  Color.LightGray,
-            trackColor =  Color.Red, //remaining health
+            color = Color.LightGray,
+            trackColor = Color.Red, //remaining health
         )
 
-        // TODO: Remove the 1- when merged to the main branch
         LinearProgressIndicator(
-            progress = { 1 - (health / 100f) },
+            progress = { (health / 100f) },
             modifier = modifier
                 .fillMaxWidth()
                 .height(15.dp),
-            color = Color.Red,
-            trackColor = Color.LightGray, //remaining health
+            color = Color.Red, //remaining health
+            trackColor = Color.LightGray,
         )
     }
 }
